refactor(bh-mod): rename validate module to check
Rename `mod validate` to `mod check` and move `validate.rs` to `check.rs`.
The entry point is `check::field`, so the module name now reads the same as
the call site.
